[IMP] core: split tests/common
Accumulation of stuff had ended up with the file being a 3KLOC mess of
stuff, some of which had no business being used (outside of
tests/common).

The addition of the WSGI stuff and rejiggering of some inheritance
hierarchies provides the occasion for a reshuffle, as it adds a fair
bit of mostly implementation details.

- all of the "public" objects are re-exported from
  `odoo.tests.common`, client code should not be impacted
- utilities which are used both internally and (maybe) externally is
  moved to .utils, small utilities which are only for external
  consumption remain at the root (for now?)
- custom test suite and test cases, chrome (CDP), and test form are
  moved into their own modules
- the browser stuff is not exposed as it should only be used
  internally by the HTTP test case (?)
- the test suite maybe should be moved laterally as it ought only be
  used by the test runner and whoever calls that?
- maybe TagsSelector should be moved somewhere else as it's also a
  runner concern?
- should the "public utilities" be moved to `odoo.tests` directly and
  `odoo.tests.common` deprecated?

While at it:

- added a utility to check for test-mode, as
  `getattr(threading.current_thread(), 'testing', False)` is an
  exceedingly common task
- moved garbage dummy files from test_assetsbundle over there because
  why the fuck?
