+++ addons/website_sale_autocomplete/tests/test_ui.py
@@ -1,7 +1,9 @@
 # -*- encoding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
+from unittest.mock import patch
+
 from odoo.addons.website_sale_autocomplete.controllers.main import AutoCompleteController
-from odoo.tests import patch, HttpCase, tagged
+from odoo.tests import HttpCase, tagged
 
 CONTROLLER_PATH = 'odoo.addons.website_sale_autocomplete.controllers.main.AutoCompleteController'
 MOCK_GOOGLE_ID = 'aHR0cHM6Ly93d3cueW91dHViZS5jb20vd2F0Y2g/dj1kUXc0dzlXZ1hjUQ=='
