+++ odoo/addons/base/tests/__init__.py
@@ -47,7 +47,6 @@ from . import test_res_partner_bank
 from . import test_res_users
 from . import test_reports
 from . import test_test_retry
-from . import test_test_suite
 from . import test_tests_tags
 from . import test_form_create
 from . import test_cloc
