+++ odoo/addons/base/tests/test_http_case.py
@@ -12,7 +12,7 @@ class TestHttpCase(HttpCase):
         with self.assertLogs(level='ERROR') as log_catcher:
             with self.assertRaises(AssertionError) as error_catcher:
                 code = "console.error('test error','message')"
-                with patch('odoo.tests.common.ChromeBrowser.take_screenshot', return_value=None):
+                with patch('odoo.tests.common.cdp.ChromeBrowser.take_screenshot', return_value=None):
                     self.browser_js(url_path='about:blank', code=code)
             # second line must contains error message
             self.assertEqual(error_catcher.exception.args[0].splitlines()[-1], "test error message")
@@ -23,7 +23,7 @@ class TestHttpCase(HttpCase):
         with self.assertLogs(level='ERROR') as log_catcher:
             with self.assertRaises(AssertionError) as error_catcher:
                 code = "console.error(TypeError('test error message'))"
-                with patch('odoo.tests.common.ChromeBrowser.take_screenshot', return_value=None):
+                with patch('odoo.tests.common.cdp.ChromeBrowser.take_screenshot', return_value=None):
                     self.browser_js(url_path='about:blank', code=code)
             # second line must contains error message
             self.assertEqual(error_catcher.exception.args[0].splitlines()[-2:],
