+++ odoo/addons/test_assetsbundle/tests/dummy.js
